Employees/Groups/Projects Changes
1.add validation against same named employees/groups/projects
2.simplified folder structure
3.assign/remove groups from projects
4.assign/remove employee from managing projects
5.altered viewing of the nav bar for anyone that is manager/admin
6.added documentation on anything public
7.moved away from seperate files for modals and controllers, instead we load them into the parent page
