+++ code/CapstoneProjectG7/TicketSystemWeb/Controllers/AccountController.cs
@@ -5,22 +5,40 @@ using TicketSystemWeb.Models.Employee;
 
 namespace TicketSystemWeb.Controllers
 {
+    /// <summary>
+    /// the account controller
+    /// </summary>
+    /// <seealso cref="Microsoft.AspNetCore.Mvc.Controller" />
     public class AccountController : Controller
     {
         private readonly SignInManager<Employee> _signInManager;
         private readonly UserManager<Employee> _userManager;
 
+        /// <summary>
+        /// Initializes a new instance of the <see cref="AccountController"/> class.
+        /// </summary>
+        /// <param name="signInManager">The sign in manager.</param>
+        /// <param name="userManager">The user manager.</param>
         public AccountController(SignInManager<Employee> signInManager, UserManager<Employee> userManager)
         {
             _signInManager = signInManager;
             _userManager = userManager;
         }
 
+        /// <summary>
+        /// Logins this instance.
+        /// </summary>
+        /// <returns>the login view</returns>
         public IActionResult Login()
         {
             return View();
         }
 
+        /// <summary>
+        /// Logins the specified model.
+        /// </summary>
+        /// <param name="model">The model.</param>
+        /// <returns>to the home page</returns>
         [HttpPost]
         public async Task<IActionResult> Login(LoginViewModel model)
         {
@@ -36,6 +54,10 @@ namespace TicketSystemWeb.Controllers
             return View(model);
         }
 
+        /// <summary>
+        /// Logouts this instance.
+        /// </summary>
+        /// <returns>back to the login view</returns>
         public async Task<IActionResult> Logout()
         {
             await _signInManager.SignOutAsync();
