+++ code/CapstoneProjectG7/TicketSystemWeb/Models/Employee/EditEmployeeViewModel.cs
@@ -1,10 +1,37 @@
 ﻿namespace TicketSystemWeb.ViewModels
 {
+    /// <summary>
+    /// editemployee viewmodel
+    /// </summary>
     public class EditEmployeeViewModel
     {
+        /// <summary>
+        /// Gets or sets the identifier.
+        /// </summary>
+        /// <value>
+        /// The identifier.
+        /// </value>
         public string Id { get; set; }
+        /// <summary>
+        /// Gets or sets the name of the user.
+        /// </summary>
+        /// <value>
+        /// The name of the user.
+        /// </value>
         public string UserName { get; set; } = string.Empty;
+        /// <summary>
+        /// Gets or sets the email.
+        /// </summary>
+        /// <value>
+        /// The email.
+        /// </value>
         public string Email { get; set; } = string.Empty;
+        /// <summary>
+        /// Gets or sets the role.
+        /// </summary>
+        /// <value>
+        /// The role.
+        /// </value>
         public string Role { get; set; } = "user";
     }
 }
