+++ code/CapstoneProjectG7/TicketSystemWeb/Models/Employee/Employee.cs
@@ -2,8 +2,18 @@
 
 namespace TicketSystemWeb.Models.Employee
 {
+    /// <summary>
+    /// employee class
+    /// </summary>
+    /// <seealso cref="Microsoft.AspNetCore.Identity.IdentityUser" />
     public class Employee : IdentityUser
     {
+        /// <summary>
+        /// Gets or sets the employee groups.
+        /// </summary>
+        /// <value>
+        /// The employee groups.
+        /// </value>
         public List<EmployeeGroup> EmployeeGroups { get; set; } = new();
     }
