+++ code/CapstoneProjectG7/TicketSystemWeb/Models/Ticket.cs
@@ -2,29 +2,74 @@
 
 namespace TicketSystemWeb.Models
 {
+    /// <summary>
+    /// the ticket class
+    /// </summary>
     public class Ticket
     {
-        [Key]  // Marks this as the Primary Key
+        /// <summary>
+        /// Gets or sets the ticket identifier.
+        /// </summary>
+        /// <value>
+        /// The ticket identifier.
+        /// </value>
+        [Key]
         public int TicketId { get; set; }
 
-        [Required]  // Ensures Title is not null
-        [StringLength(100)]  // Limits Title to 100 characters
+        /// <summary>
+        /// Gets or sets the title.
+        /// </summary>
+        /// <value>
+        /// The title.
+        /// </value>
+        [Required]
+        [StringLength(100)]
         public string Title { get; set; }
 
+        /// <summary>
+        /// Gets or sets the description.
+        /// </summary>
+        /// <value>
+        /// The description.
+        /// </value>
         [Required]
-        [StringLength(500)]  // Description can be up to 500 characters
+        [StringLength(500)]
         public string Description { get; set; }
 
+        /// <summary>
+        /// Gets or sets the status.
+        /// </summary>
+        /// <value>
+        /// The status.
+        /// </value>
         [Required]
-        public string Status { get; set; } // Example: "Open", "In Progress", "Closed"
+        public string Status { get; set; }
 
+        /// <summary>
+        /// Gets or sets the created at.
+        /// </summary>
+        /// <value>
+        /// The created at.
+        /// </value>
         [Required]
-        public DateTime CreatedAt { get; set; } = DateTime.UtcNow;  // Default to current time
+        public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
 
-        public DateTime? ClosedAt { get; set; }  // Nullable if ticket is not closed yet
+        /// <summary>
+        /// Gets or sets the closed at.
+        /// </summary>
+        /// <value>
+        /// The closed at.
+        /// </value>
+        public DateTime? ClosedAt { get; set; }
 
+        /// <summary>
+        /// Gets or sets the created by.
+        /// </summary>
+        /// <value>
+        /// The created by.
+        /// </value>
         [Required]
-        public string CreatedBy { get; set; }  // The user who created the ticket
+        public string CreatedBy { get; set; }
     }
 
 }
