--- code/CapstoneProjectG7/TicketSystemWeb/Views/ProjectManagement/Groups/EditGroup.cshtml
@@ -1,49 +1,0 @@
-﻿@model TicketSystemWeb.ViewModels.EditGroupViewModel
-
-<div class="modal fade" id="editGroupModal" tabindex="-1" aria-labelledby="editGroupModalLabel" aria-hidden="true">
-    <div class="modal-dialog">
-        <div class="modal-content">
-            <div class="modal-header">
-                <h5 class="modal-title" id="editGroupModalLabel">Edit Group</h5>
-                <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
-            </div>
-            <form asp-controller="EditGroup" asp-action="UpdateGroup" method="post">
-                <input type="hidden" id="editGroupId" name="GroupId" value="@Model.GroupId" />
-                <div class="modal-body">
-                    <div class="mb-3">
-                        <label for="editGroupName" class="form-label">Group Name</label>
-                        <input type="text" class="form-control" id="editGroupName" name="Name" value="@Model.Name" required>
-                    </div>
-                    <div class="mb-3">
-                        <label for="editManager" class="form-label">Assign Manager</label>
-                        <select class="form-control" id="editManager" name="SelectedManagerId" required>
-                            <option value="">-- Select Manager --</option>
-                            @foreach (var employee in Model.AllEmployees)
-                            {
-                                <option value="@employee.Id" selected="@(employee.Id == Model.SelectedManagerId ? "selected" : null)">
-                                    @employee.UserName
-                                </option>
-                            }
-                        </select>
-                    </div>
-                    <div class="mb-3">
-                        <label for="editEmployees" class="form-label">Add Employees</label>
-                        <select multiple class="form-control" id="editEmployees" name="SelectedEmployeeIds">
-                            @foreach (var employee in Model.AllEmployees)
-                            {
-                                <option value="@employee.Id" selected="@(Model.SelectedEmployeeIds.Contains(employee.Id) ? "selected" : null)">
-                                    @employee.UserName
-                                </option>
-                            }
-                        </select>
-                        <small class="form-text text-muted">Hold Ctrl (Cmd on Mac) to select multiple employees.</small>
-                    </div>
-                </div>
-                <div class="modal-footer">
-                    <button type="submit" class="btn btn-success" style="background-color: #6f42c1; border-color: #6f42c1; color: white;">Save Changes</button>
-                    <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Cancel</button>
-                </div>
-            </form>
-        </div>
-    </div>
-</div>
