--- code/CapstoneProjectG7/TicketSystemWeb/Views/ProjectManagement/Groups/RemoveGroup.cshtml
@@ -1,22 +1,0 @@
-﻿@model int
-
-<div class="modal fade" id="removeGroupModal" tabindex="-1" aria-labelledby="removeGroupModalLabel" aria-hidden="true">
-    <div class="modal-dialog">
-        <div class="modal-content">
-            <div class="modal-header">
-                <h5 class="modal-title" id="removeGroupModalLabel">Confirm Group Removal</h5>
-                <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
-            </div>
-            <form asp-controller="RemoveGroup" asp-action="DeleteGroup" method="post">
-                <input type="hidden" id="removeGroupId" name="groupId" value="@Model" />
-                <div class="modal-body">
-                    <p>Are you sure you want to remove this group?</p>
-                </div>
-                <div class="modal-footer">
-                    <button type="submit" class="btn btn-danger">Remove</button>
-                    <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Cancel</button>
-                </div>
-            </form>
-        </div>
-    </div>
-</div>
