--- code/CapstoneProjectG7/TicketSystemWeb/Views/ProjectManagement/Projects/AddProject.cshtml
@@ -1,37 +1,0 @@
-﻿@model TicketSystemWeb.ViewModels.AddProjectViewModel
-
-<div class="modal fade" id="addProjectModal" tabindex="-1" aria-labelledby="addProjectModalLabel" aria-hidden="true">
-    <div class="modal-dialog">
-        <div class="modal-content">
-            <div class="modal-header">
-                <h5 class="modal-title" id="addProjectModalLabel">Add Project</h5>
-                <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
-            </div>
-            <form asp-controller="AddProject" asp-action="Create" method="post">
-                <div class="modal-body">
-                    <div class="mb-3">
-                        <label for="title" class="form-label">Title</label>
-                        <input type="text" class="form-control" name="Title" required>
-                    </div>
-                    <div class="mb-3">
-                        <label for="description" class="form-label">Description</label>
-                        <textarea class="form-control" name="Description" required></textarea>
-                    </div>
-                    <div class="mb-3">
-                        <label for="groups" class="form-label">Assign Groups</label>
-                        <select multiple class="form-control" name="GroupIds">
-                            @foreach (var group in Model.AllGroups)
-                            {
-                                <option value="@group.Id">@group.Name</option>
-                            }
-                        </select>
-                    </div>
-                </div>
-                <div class="modal-footer">
-                    <button type="submit" class="btn btn-primary" style="background-color: #6f42c1; border-color: #6f42c1; color: white;">Add</button>
-                    <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Cancel</button>
-                </div>
-            </form>
-        </div>
-    </div>
-</div>
