--- code/CapstoneProjectG7/TicketSystemWeb/Views/ProjectManagement/Projects/EditProject.cshtml
@@ -1,41 +1,0 @@
-﻿@model TicketSystemWeb.ViewModels.EditProjectViewModel
-
-<div class="modal fade" id="editProjectModal" tabindex="-1" aria-labelledby="editProjectModalLabel" aria-hidden="true">
-    <div class="modal-dialog">
-        <div class="modal-content">
-            <div class="modal-header">
-                <h5 class="modal-title" id="editProjectModalLabel">Edit Project</h5>
-                <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
-            </div>
-            <form asp-controller="EditProject" asp-action="UpdateProject" method="post">
-                <input type="hidden" id="editProjectId" name="ProjectId" value="" />
-                <div class="modal-body">
-                    <div class="mb-3">
-                        <label for="editTitle" class="form-label">Title</label>
-                        <input type="text" class="form-control" id="editTitle" name="Title" required>
-                    </div>
-                    <div class="mb-3">
-                        <label for="editDescription" class="form-label">Description</label>
-                        <textarea class="form-control" id="editDescription" name="Description" required></textarea>
-                    </div>
-                    <div class="mb-3">
-                        <label for="editGroups" class="form-label">Assign Groups</label>
-                        <select multiple class="form-control" id="editGroups" name="SelectedGroupIds">
-                            @foreach (var group in Model.AllGroups)
-                            {
-                                <option value="@group.Id" selected="@(Model.SelectedGroupIds.Contains(group.Id) ? "selected" : null)">
-                                    @group.Name
-                                </option>
-                            }
-                        </select>
-                        <small class="form-text text-muted">Hold Ctrl (Cmd on Mac) to select multiple groups.</small>
-                    </div>
-                </div>
-                <div class="modal-footer">
-                    <button type="submit" class="btn btn-success" style="background-color: #6f42c1; border-color: #6f42c1; color: white;">Save Changes</button>
-                    <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Cancel</button>
-                </div>
-            </form>
-        </div>
-    </div>
-</div>
