--- code/CapstoneProjectG7/TicketSystemWeb/Views/ProjectManagement/Projects/RemoveProject.cshtml
@@ -1,22 +1,0 @@
-﻿@model int
-
-<div class="modal fade" id="removeProjectModal" tabindex="-1" aria-labelledby="removeProjectModalLabel" aria-hidden="true">
-    <div class="modal-dialog">
-        <div class="modal-content">
-            <div class="modal-header">
-                <h5 class="modal-title" id="removeProjectModalLabel">Confirm Project Removal</h5>
-                <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
-            </div>
-            <form asp-controller="RemoveProject" asp-action="DeleteProject" method="post">
-                <input type="hidden" id="removeProjectId" name="projectId" value="" />
-                <div class="modal-body">
-                    <p>Are you sure you want to remove this project?</p>
-                </div>
-                <div class="modal-footer">
-                    <button type="submit" class="btn btn-danger">Remove</button>
-                    <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Cancel</button>
-                </div>
-            </form>
-        </div>
-    </div>
-</div>
